+++ darwin-web/src/main/java/xin/manong/darwin/web/controller/RuleGroupController.java
@@ -86,7 +86,7 @@ public class RuleGroupController {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("get")
     @GetMapping("get")
-    public RuleGroup get(@QueryParam("id") Long id) {
+    public RuleGroup get(@QueryParam("id") Integer id) {
         if (id == null) {
             logger.error("missing param[id]");
             throw new BadRequestException("规则分组ID缺失");
@@ -150,7 +150,7 @@ public class RuleGroupController {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("delete")
     @DeleteMapping("delete")
-    public Boolean delete(@QueryParam("id") Long id) {
+    public Boolean delete(@QueryParam("id") Integer id) {
         if (id == null) {
             logger.error("missing param[id]");
             throw new BadRequestException("规则分组ID缺失");
